requirements.txt
bcrypt==4.0.1
cffi==1.15.1
cryptography==41.0.3
paramiko==3.3.1
pycparser==2.21
PyNaCl==1.5.0
python-dotenv==1.0.0
RouterOS-api==0.17.0
six==1.16.0
